update project
add new classes
add HandlerException
add Json annotations

+++ src/main/java/com/gicci/playground/service/CrudService.java
@@ -9,8 +9,8 @@ public interface CrudService<T> {
 
 	public T create(T type) throws RecordConstraint;
 	public T update(T type) throws RecordNotFound;
-	public T delete(Integer id) throws RecordNotFound;
-	public T findById(Integer id) throws RecordNotFound;
+	public T delete(Long id) throws RecordNotFound;
+	public T findById(Long id) throws RecordNotFound;
 	public List<T> findAll();
 	
 }
